fix(ci): file-line-limit — split executor forge-gate tests into part_015
test_executor_error_paths_part_007.py crossed the 1,500-line limit after this
PR added two sync_release_pr forge-gate tests. Move them into a new sibling part
(test_executor_error_paths_part_015.py) that imports the shared release-sync
scaffolding from part_007, matching how parts 008-013 reuse base parts. No test
behavior changes; both files now sit comfortably under the limit.
